sakura: Extension to R Serialization

Extends the functionality of R serialization by augmenting the built-in reference hook system. This enhanced implementation allows an integrated single-pass operation that combines R serialization with third-party serialization methods. Facilitates the serialization of even complex R objects, which contain non-system reference objects, such as those accessed via external pointers, to enable their use in parallel and distributed computing.

See also

Author

Maintainer: Charlie Gao charlie.gao@shikokuchuo.net (ORCID)

Authors:

Travers Ching traversc@gmail.com

Other contributors:

Posit Software, PBC [copyright holder, funder]

Hibiki AI Limited [copyright holder]